refactor storage of models
