--- src-tauri/src/filesystem/infos/mod.rs
@@ -1,1 +1,0 @@
-pub mod volume_information;
