+++ src-tauri/src/filesystem/models/mod.rs
@@ -2,3 +2,6 @@ mod directory;
 pub use directory::Directory;
 mod file;
 pub use file::File;
+mod volume;
+pub use volume::VolumeInformation;
+
